[add]Bookコントローラー処理時の認可処理

+++ app/Http/Controllers/BookController.php
@@ -98,6 +98,7 @@ class BookController extends Controller
 
     public function store(Book $book, BookRequest $request)
     {
+
         if(is_uploaded_file($_FILES['cover']['tmp_name'])){
             $upload_image = $request->file('cover');
             $file_name = time() . '_' . $upload_image->getClientOriginalName();
@@ -118,6 +119,8 @@ class BookController extends Controller
 
     public function show(Book $book, Request $request)
     {
+        $this->authorize('view', $book);
+
         $book = Book::find($book->id);
         $keyword = $request->keyword;
         $tag = $request->tag;
@@ -153,11 +156,15 @@ class BookController extends Controller
 
     public function edit(Book $book)
     {
+        $this->authorize('update', $book);
         return view('books.edit', compact('book'));
     }
 
     public function update(Book $book, BookRequest $request)
     {
+
+        $this->authorize('update', $book);
+
         if(is_uploaded_file($_FILES['cover']['tmp_name'])){
             $upload_image = $request->file('cover');
             $file_name = time() . '_' . $upload_image->getClientOriginalName();
@@ -179,6 +186,8 @@ class BookController extends Controller
 
     public function destroy(Book $book)
     {
+        $this->authorize('update', $book);
+
         $book->memos()->each(function ($memo) {
             $memo->delete();
         });
